khepri_app (khepri v0.19.0)

The Khepri application

Khepri supports the following application environment variables:

default_ra_system: the name of the Ra system by default for Khepri stores.

default_store_id: the name of the default store, used by API that don't take a StoreId.

default_timeout: the default timeout.

Functions

config_change(Changed, New, Removed)

get_default_timeout()

start(_, _)

stop(_)