khepri_event_handler (khepri v0.19.0)

Khepri event handling server.

Types

trigger_action/0

-type trigger_action() ::
          {sproc, khepri_path:native_path()} |
          {apply, {module(), atom(), list()}} |
          {send, pid(), any()}.

The type of action associated with a trigger.

It can be the following actions:

sproc: the stored procedure pointed to by the given path will be executed

apply, the function described by the given module/function/arguments tuple will be executed

send: a message will be sent to the specified PID with the event details and the given term

The action will get a trigger desccriptor as argument to describe a specific execution of that trigger. See trigger_descriptor/0.

trigger_descriptor/0

-type trigger_descriptor() ::
          #khepri_trigger{type :: tree | process,
                          store_id :: khepri:store_id(),
                          trigger_id :: khepri:trigger_id(),
                          event :: map(),
                          action :: map()}.

Descriptor of the trigger used as an argument to the trigger action.

The descriptor contains all the properties of a specific instance of that trigger.

When the action is a stored procedure or an MFA tuple, the descriptor is appended to the list of arguments.

When the action is a PID, the descriptor is send to the PID as a standalone message. The action properties map will contain the Priv element of the send tuple under the priv field.

trigger_exec_loc/0

-type trigger_exec_loc() :: leader | {member, node()} | all_members.

Where to execute the triggered action.

It supports the following locations:

leader: the action is executed on the leader node at the time the action is triggered.

{member, Member}: the action is executed on the given node if it is a member of the cluster, or the leader otherwise.

all_members: the action is executed on all members of the cluster.

triggered_action/0

-type triggered_action() ::
          {sproc, horus:horus_fun()} | {apply, {module(), atom(), list()}} | {send, pid(), any()}.

The action associated with a trigger once it is triggered.

It can be the following actions:

sproc: the stored procedure (extracted from the path given to the trigger registration) is executed

apply: the function described by the given module/function/arguments tuple is executed

send: a message is sent to the specified PID with the event details and the given term

Functions

code_change(OldVsn, State, Extra)

handle_call(Request, From, State)

handle_cast(Request, State)

handle_info(Msg, State)

handle_triggered_actions(StoreId, TriggeredActions)

init(_)

start_link()

terminate(Reason, State)