khepri_machine (khepri v0.4.3)

Khepri private low-level API.

This module exposes the private "low-level" API to the Khepri database and state machine. Main functions correspond to Ra commands implemented by the state machine. All functions in khepri are built on top of this module.
This module is private. The documentation is still visible because it may help understand some implementation details. However, this module should never be called directly outside of Khepri.

Types

async_ret/0

-type async_ret() :: ok.

command/0

-type command() :: #put{} | #delete{} | #tx{} | #register_trigger{} | #ack_triggered{}.

Commands specific to this Ra machine.

common_ret/0

-type common_ret() :: khepri:ok(khepri_adv:node_props_map()) | khepri:error().

keep_while_aftermath/0

-type keep_while_aftermath() :: #{khepri_path:native_path() => khepri:node_props() | delete}.

Internal index of the per-node changes which happened during a traversal. This is used when the tree is walked back up to determine the list of tree nodes to remove after some keep_while condition evaluates to false.

keep_while_conds_map/0

-type keep_while_conds_map() :: #{khepri_path:native_path() => khepri_condition:native_keep_while()}.

Per-node keep_while conditions.

keep_while_conds_revidx/0

-type keep_while_conds_revidx() :: #{khepri_path:native_path() => #{khepri_path:native_path() => ok}}.

Internal reverse index of the keep_while conditions. If node A depends on a condition on node B, then this reverse index will have a "node B => node A" entry.

machine_config/0

-type machine_config() :: #config{}.

Configuration record, holding read-only or rarely changing fields.

machine_init_args/0

-type machine_init_args() ::
    #{store_id := khepri:store_id(),
      member := ra:server_id(),
      snapshot_interval => non_neg_integer(),
      commands => [command()],
      atom() => any()}.

Structure passed to init/1.

ok/2

-type ok(Type1, Type2) :: {ok, Type1, Type2}.

ok/3

-type ok(Type1, Type2, Type3) :: {ok, Type1, Type2, Type3}.

props/0

-type props() ::
    #{payload_version := khepri:payload_version(),
      child_list_version := khepri:child_list_version()}.

Properties attached to each node in the tree structure.

query_fun/0

-type query_fun() :: fun((state()) -> any()).

Function representing a query and used process_query/3.

state/0

-type state() :: #khepri_machine{}.

State of this Ra state machine.

tree_node/0

-type tree_node() :: #node{}.

A node in the tree structure.

triggered/0

-type triggered() :: #triggered{}.

tx_ret/0

-type tx_ret() :: khepri:ok(khepri_tx:tx_fun_result()) | khepri_tx:tx_abort() | no_return().

walk_down_the_tree_extra/0

-type walk_down_the_tree_extra() ::
    #{include_root_props => boolean(),
      keep_while_conds => keep_while_conds_map(),
      keep_while_conds_revidx => keep_while_conds_revidx(),
      keep_while_aftermath => keep_while_aftermath()}.

walk_down_the_tree_fun/0

-type walk_down_the_tree_fun() ::
    fun((khepri_path:native_path(), tree_node() | {interrupted, any(), map()}, Acc :: any()) ->
            ok(tree_node() | keep | delete, any()) | khepri:error()).

Function called to handle a node found (or an error) and used in walk_down_the_tree/6.

Functions

count(StoreId, PathPattern, Options)

-spec count(StoreId, PathPattern, Options) -> Ret
         when
             StoreId :: khepri:store_id(),
             PathPattern :: khepri_path:pattern(),
             Options :: khepri:query_options() | khepri:tree_options(),
             Ret :: khepri:ok(Count) | khepri:error(),
             Count :: non_neg_integer().

Counts all tree nodes matching the path pattern.

delete(StoreId, PathPattern, Options)

-spec delete(StoreId, PathPattern, Options) -> Ret
          when
              StoreId :: khepri:store_id(),
              PathPattern :: khepri_path:pattern(),
              Options :: khepri:command_options() | khepri:tree_options(),
              Ret :: khepri_machine:common_ret() | khepri_machine:async_ret().

Deletes all tree nodes matching the path pattern.

get(StoreId, PathPattern, Options)

-spec get(StoreId, PathPattern, Options) -> Ret
       when
           StoreId :: khepri:store_id(),
           PathPattern :: khepri_path:pattern(),
           Options :: khepri:query_options() | khepri:tree_options(),
           Ret :: khepri_machine:common_ret().

Returns all tree nodes matching the given path pattern.

handle_tx_exception(_)

register_trigger(StoreId, TriggerId, EventFilter, StoredProcPath, Options)

-spec register_trigger(StoreId, TriggerId, EventFilter, StoredProcPath, Options) -> Ret
                    when
                        StoreId :: khepri:store_id(),
                        TriggerId :: khepri:trigger_id(),
                        EventFilter :: khepri_evf:event_filter() | khepri_path:pattern(),
                        StoredProcPath :: khepri_path:path(),
                        Options :: khepri:command_options(),
                        Ret :: ok | khepri:error().

Registers a trigger.

run_sproc(StoreId, PathPattern, Args, Options)

-spec run_sproc(StoreId, PathPattern, Args, Options) -> Ret
             when
                 StoreId :: khepri:store_id(),
                 PathPattern :: khepri_path:pattern(),
                 Args :: [any()],
                 Options :: khepri:query_options(),
                 Ret :: any().

Executes a stored procedure.
The stored procedure is executed in the context of the caller of run_sproc/3.

transaction(StoreId, Fun, ReadWrite, Options)

-spec transaction(StoreId, Fun, ReadWrite, Options) -> Ret
               when
                   StoreId :: khepri:store_id(),
                   Fun :: khepri_tx:tx_fun(),
                   ReadWrite :: ro | rw | auto,
                   Options :: khepri:command_options() | khepri:query_options(),
                   Ret :: khepri_machine:tx_ret() | khepri_machine:async_ret().

Runs a transaction and returns the result.