khepri_prefix_tree (khepri v0.19.0)

Types

tree/1

-opaque tree(Payload)

Functions

empty()

-spec empty() -> tree(_).

Returns a new empty tree.

See also: tree().

find_path(Path, Tree)

-spec find_path(Path, Tree) -> Ret
                   when
                       Path :: khepri_path:native_path(),
                       Tree :: khepri_prefix_tree:tree(Payload),
                       Payload :: any(),
                       Ret :: {ok, Payload} | error.

Returns the payload associated with a path in the tree.

fold_prefixes_of(Fun, Acc, Path, Tree)

-spec fold_prefixes_of(Fun, Acc, Path, Tree) -> Ret
                          when
                              Fun :: fun((Payload, Acc) -> Acc1),
                              Acc :: any(),
                              Acc1 :: any(),
                              Path :: khepri_path:native_path(),
                              Tree :: khepri_prefix_tree:tree(Payload),
                              Payload :: any(),
                              Ret :: Acc1.

Folds over all nodes in the tree which are prefixed by the given Path building an accumulated value with the given fold function and initial accumulator.

from_map(Map)

-spec from_map(Map) -> Tree
                  when
                      Map :: #{khepri_path:native_path() => Payload},
                      Tree :: khepri_prefix_tree:tree(Payload),
                      Payload :: any().

Converts a map of paths to payloads into a prefix tree.

is_prefix_tree(Prefix_tree)

-spec is_prefix_tree(tree(_)) -> true;
                    (any()) -> false.

Determines whether the given term is a prefix tree.

update(Fun, Path, Tree)

-spec update(Fun, Path, Tree) -> Ret
                when
                    Fun :: fun((Payload | '$__NO_PAYLOAD__') -> Payload | '$__NO_PAYLOAD__'),
                    Path :: khepri_path:native_path(),
                    Tree :: khepri_prefix_tree:tree(Payload),
                    Payload :: any(),
                    Ret :: khepri_prefix_tree:tree(Payload).

Updates a given path in the tree.

This function can be used to create, update or delete tree nodes. If the tree node does not exist for the given path, the update function is passed ?NO_PAYLOAD. If the update function returns ?NO_PAYLOAD then the tree node and all of its ancestors which do not have a payload or children are removed.

The update function is also be passed ?NO_PAYLOAD if a tree node exists but does not have a payload: being passed ?NO_PAYLOAD is not a reliable sign that a tree node did not exist prior to an update.