khepri_machine (khepri v0.19.0)

Khepri private low-level API.

This module exposes the private "low-level" API to the Khepri database and state machine. Main functions correspond to Ra commands implemented by the state machine. All functions in khepri are built on top of this module.

This module is private. The documentation is still visible because it may help understand some implementation details. However, this module should never be called directly outside of Khepri.

State machine history

<table><tr><th>Version</th><th>What changed</th></tr><tr><td>0</td><td>Initial version</td></tr><tr><td>1</td><td>

Added deduplication mechanism:

new command option protect_against_dups

new commands #dedup{} and #dedup_ack{}

new state field dedups

Added command #unregister_projections{}. The previous #unregister_projection{} is still supported for backward-compatibility but it is no longer created.

</td></tr><tr><td>2</td><td>

Changed the data structure for the reverse index used to track keep-while conditions to be a prefix tree (see khepri_prefix_tree).

Moved the expiration of dedups to the tick aux effect. This also introduces a new command #drop_dedups{}.

Added the delete_reason to the list of properties that can be returned. It is returned by default if the effective machine version is 2 or more.

</td></tr><tr><td>3</td><td>

Added support for multi-table projections (see khepri_projection).

</td></tr><tr><td>4</td><td>

Added uniform commands, replacing all existing commands.

Added #request_snapshot{} helper command for tests.

Started to record init arguments in the machine state.

Started to track unreleased command footprint in state machine metrics.

Started to use unreleased command footprint to determine if a snapshot is needed.

Started to use the last snasphot timestamp to determine if a snapshot is needed.

Extended the trigger mechanism to support other types of action, in addition to stored procedures. This comes with two new records: the #register_trigger_v2{} command and #triggered_v2{}, stored in the machine state. The extended mechanism continues to support stored procedure (the previous mechanism is not used with version 4).

Started to cache the list of cluster member nodenames in the machine state. The machine state was modified to add the cached_members list at its end.

Added the where option for triggers, to allow the caller to decide on which cluster member(s) the trigger action is executed. See khepri:trigger_options().

Added support for process-lifetime-based triggers. The trigger can be executed when a process exits.

Added support for process-lifetime-based keep_while condition. The keep_while put option can take a PID. The tree node will be deleted as soon as the associated process exits. See khepri_condition:keep_while().

Added support for "apply MFA" and "send message" trigger actions.

</td></tr></table>

Types

api_behaviour/0

-type api_behaviour() ::
          dedup_protection | delete_reason_in_node_props | expire_dedups_from_tick |
          indirect_deletes_in_ret | uniform_write_ret | multi_table_projections | uniform_commands |
          request_snapshot | extended_trigger | cached_members_list | process_based_keep_while.

Name of a state machine API behaviour.

async_ret/0

-type async_ret() :: ok.

aux_query/0

-type aux_query() :: {query, query_fun(), khepri:query_options()}.

Query executed through the Ra aux mechanism.

cached_members_list/0

-type cached_members_list() :: [node() | node()].

Non-empty list of cluster member nodenames.

command/0

-type command() ::
          #put_v{args ::
                     #put_v1{path :: khepri_path:native_pattern(),
                             payload :: khepri_payload:payload(),
                             put_options :: khepri:put_options(),
                             tree_options :: khepri:tree_options()},
                 common ::
                     #common_v1{command_size :: non_neg_integer(),
                                dedup_ref :: reference() | undefined,
                                dedup_expiry :: integer() | undefined} |
                     none} |
          #delete_v{args ::
                        #delete_v1{path :: khepri_path:native_pattern(),
                                   tree_options :: khepri:tree_options()},
                    common ::
                        #common_v1{command_size :: non_neg_integer(),
                                   dedup_ref :: reference() | undefined,
                                   dedup_expiry :: integer() | undefined} |
                        none} |
          #tx_v{args :: #tx_v1{'fun' :: horus:horus_fun() | khepri_path:pattern(), args :: list()},
                common ::
                    #common_v1{command_size :: non_neg_integer(),
                               dedup_ref :: reference() | undefined,
                               dedup_expiry :: integer() | undefined} |
                    none} |
          #register_trigger_v{args ::
                                  #register_trigger_v1{id :: khepri:trigger_id(),
                                                       event_filter :: khepri_evf:event_filter(),
                                                       action :: khepri_event_handler:trigger_action(),
                                                       options :: khepri:trigger_options()},
                              common ::
                                  #common_v1{command_size :: non_neg_integer(),
                                             dedup_ref :: reference() | undefined,
                                             dedup_expiry :: integer() | undefined} |
                                  none} |
          #ack_triggered_v{args :: #ack_triggered_v1{triggered :: [khepri_machine:triggered()]},
                           common ::
                               #common_v1{command_size :: non_neg_integer(),
                                          dedup_ref :: reference() | undefined,
                                          dedup_expiry :: integer() | undefined} |
                               none} |
          #register_projection_v{args ::
                                     #register_projection_v1{pattern :: khepri_path:native_pattern(),
                                                             projection ::
                                                                 khepri_projection:projection()},
                                 common ::
                                     #common_v1{command_size :: non_neg_integer(),
                                                dedup_ref :: reference() | undefined,
                                                dedup_expiry :: integer() | undefined} |
                                     none} |
          #unregister_projections_v{args ::
                                        #unregister_projections_v1{names ::
                                                                       all | [khepri_projection:name()]},
                                    common ::
                                        #common_v1{command_size :: non_neg_integer(),
                                                   dedup_ref :: reference() | undefined,
                                                   dedup_expiry :: integer() | undefined} |
                                        none} |
          #dedup_ack_v{args :: #dedup_ack_v1{ref :: reference()},
                       common ::
                           #common_v1{command_size :: non_neg_integer(),
                                      dedup_ref :: reference() | undefined,
                                      dedup_expiry :: integer() | undefined} |
                           none} |
          #drop_dedups_v{args :: #drop_dedups_v1{refs :: [reference()]},
                         common ::
                             #common_v1{command_size :: non_neg_integer(),
                                        dedup_ref :: reference() | undefined,
                                        dedup_expiry :: integer() | undefined} |
                             none} |
          #request_snapshot{args :: #request_snapshot_v1{reason :: string()},
                            common ::
                                #common_v1{command_size :: non_neg_integer(),
                                           dedup_ref :: reference() | undefined,
                                           dedup_expiry :: integer() | undefined} |
                                none} |
          #cache_members_list{args ::
                                  #cache_members_list_v1{members :: khepri_machine:cached_members_list()},
                              common ::
                                  #common_v1{command_size :: non_neg_integer(),
                                             dedup_ref :: reference() | undefined,
                                             dedup_expiry :: integer() | undefined} |
                                  none}.

Commands specific to this Ra machine.

dedups_map/0

-type dedups_map() :: #{reference() => {any(), integer()}}.

Map to handle command deduplication.

delayed_aux_query/0

-type delayed_aux_query() :: {aux_query(), gen_statem:from(), ra:query_condition() | none}.

Aux query that is delayed until a condition is met.

machine_init_args/0

-type machine_init_args() ::
          #{store_id := khepri:store_id(),
            member := ra:server_id(),
            snapshot_interval => non_neg_integer(),
            commands =>
                [command() | old_command() | {machine_version, non_neg_integer(), non_neg_integer()}],
            atom() => any()}.

Structure passed to init/1.

metrics/0

-type metrics() ::
          #{applied_command_count => non_neg_integer(),
            unreleased_command_footprint => non_neg_integer(),
            last_snapshot_request_timestamp => integer(),
            init_args => khepri_machine:machine_init_args()}.

old_command/0

-type old_command() ::
          #put{path :: khepri_path:native_pattern(),
               payload :: khepri_payload:payload(),
               options :: khepri:tree_options() | khepri:put_options()} |
          #delete{path :: khepri_path:native_pattern(), options :: khepri:tree_options()} |
          #tx{'fun' :: horus:horus_fun() | khepri_path:pattern(), args :: list()} |
          #register_trigger{id :: khepri:trigger_id(),
                            event_filter :: khepri_evf:event_filter(),
                            sproc :: khepri_path:native_path()} |
          #ack_triggered{triggered :: [khepri_machine:triggered()]} |
          #register_projection{pattern :: khepri_path:native_pattern(),
                               projection :: khepri_projection:projection()} |
          #unregister_projection{name :: khepri_projection:name()} |
          #unregister_projections{names :: all | [khepri_projection:name()]} |
          #dedup{ref :: reference(),
                 expiry :: integer(),
                 command :: khepri_machine:command() | khepri_machine:old_command()} |
          #dedup_ack{ref :: reference()} |
          #drop_dedups{refs :: [reference()]}.

Old commands that are still accepted by the Ra machine but never created.

Even though Khepri no longer creates these commands, they may still be present in existing Ra log files and thus be applied after an ugprade of Khepri.

We keep them supported for backward-compatibility.

projection_map/0

-type projection_map() :: #{khepri_projection:name() => khepri_path:pattern()}.

A mapping between the names of projections and patterns to which each projection is registered.

projection_tree/0

-type projection_tree() :: khepri_pattern_tree:tree([khepri_projection:projection()]).

A pattern tree that holds all registered projections in the machine's state.

props/0

-type props() ::
          #{payload_version := khepri:payload_version(),
            child_list_version := khepri:child_list_version()}.

Properties attached to each node in the tree structure.

query_fun/0

-type query_fun() :: fun((state()) -> any()) | fun((ra_machine:command_meta_data(), state()) -> any()).

Function representing a query and used process_query/3.

state/0

-type state() :: state_v4() | khepri_machine_v1:state() | khepri_machine_v0:state().

State of this Ra state machine.

state_v4/0

-opaque state_v4()

State of this Ra state machine, version 4.

triggered/0

-type triggered() ::
          #triggered{id :: khepri:trigger_id(),
                     event_filter :: khepri_evf:event_filter(),
                     sproc :: horus:horus_fun(),
                     props :: map()}.

Record sent from khepri_event_handler to the state machine to acknowledge the execution of a trigger.

triggered_list/0

-type triggered_list() :: [khepri_machine:triggered() | khepri_machine:triggered_v2()].

List of #triggered{} or #triggered_v2{} records.

triggered_v2/0

-type triggered_v2() ::
          #triggered_v2{id :: khepri:trigger_id(),
                        event_filter :: khepri_evf:event_filter(),
                        action :: khepri_event_handler:triggered_action(),
                        where :: khepri_event_handler:trigger_exec_loc(),
                        event :: term()}.

Record sent from khepri_event_handler to the state machine to acknowledge the execution of a trigger.

It replaced triggered/0 in the state machine version 4.

triggers_map/0

-type triggers_map() ::
          #{khepri:trigger_id() =>
                #{event_filter := khepri_evf:event_filter(), sproc := khepri_path:native_path()}}.

Internal triggers map in the machine state.

triggers_map_v2/0

-type triggers_map_v2() ::
          #{khepri:trigger_id() =>
                #{event_filter := khepri_evf:event_filter(),
                  action := khepri_event_handler:trigger_action(),
                  where := khepri_event_handler:trigger_exec_loc()}}.

Internal triggers map in the machine state (v2).

This new format was introduced in version 4 of the state machine to take into account the execution of an MFA or the send of a message to a process.

tx_ret/0

-type tx_ret() :: khepri:ok(khepri_tx:tx_fun_result()) | khepri_tx:tx_abort() | no_return().

write_ret/0

-type write_ret() :: khepri:ok(khepri:node_props_map()) | khepri:error().

Functions

convert_to_uniform_command(Command)

-spec convert_to_uniform_command(Command) -> NewCommand
                                    when
                                        Command ::
                                            khepri_machine:command() | khepri_machine:old_command(),
                                        NewCommand :: khepri_machine:command().

delete(StoreId, PathPattern, Options)

-spec delete(StoreId, PathPattern, Options) -> Ret
                when
                    StoreId :: khepri:store_id(),
                    PathPattern :: khepri_path:pattern(),
                    Options :: khepri:command_options() | khepri:tree_options(),
                    Ret :: khepri_machine:write_ret() | khepri_machine:async_ret().

Deletes all tree nodes matching the path pattern.

does_api_comply_with(Behaviour, MacVer)

-spec does_api_comply_with(Behaviour, MacVer | StoreId) -> DoesUse
                              when
                                  Behaviour :: khepri_machine:api_behaviour(),
                                  MacVer :: ra_machine:version(),
                                  StoreId :: khepri:store_id(),
                                  DoesUse :: boolean().

Indicates if a new behaviour of the transaction API is activated.

The transaction code is compiled on one Erlang node with a specific version of Khepri. However, it is executed on all members of the Khepri cluster. Some Erlang nodes might use another version of Khepri, newer or older, and the transaction API may differ.

For instance in Khepri 0.17.x, the return values of the khepri_tx_adv functions changed. The transaction code will have to handle both versions of the API to work correctly. Thus it can use this function to adapt.

fence(StoreId, Timeout)

-spec fence(StoreId, Timeout) -> Ret
               when StoreId :: khepri:store_id(), Timeout :: timeout(), Ret :: ok | khepri:error().

Blocks until all updates received by the cluster leader are applied locally.

fold(StoreId, PathPattern, Fun, Acc, Options)

-spec fold(StoreId, PathPattern, Fun, Acc, Options) -> Ret
              when
                  StoreId :: khepri:store_id(),
                  PathPattern :: khepri_path:pattern(),
                  Fun :: khepri:fold_fun(),
                  Acc :: khepri:fold_acc(),
                  Options :: khepri:query_options() | khepri:tree_options(),
                  Ret :: khepri:ok(NewAcc) | khepri:error(),
                  NewAcc :: Acc.

Returns all tree nodes matching the given path pattern.

handle_tx_exception(_)

register_projection(StoreId, PathPattern, Projection, Options)

-spec register_projection(StoreId, PathPattern, Projection, Options) -> Ret
                             when
                                 StoreId :: khepri:store_id(),
                                 PathPattern :: khepri_path:pattern(),
                                 Projection :: khepri_projection:projection(),
                                 Options :: khepri:command_options(),
                                 Ret :: ok | khepri:error().

Registers a projection.

register_trigger(StoreId, TriggerId, EventFilter, Action, Options)

-spec register_trigger(StoreId, TriggerId, EventFilter, Action, Options) -> Ret
                          when
                              StoreId :: khepri:store_id(),
                              TriggerId :: khepri:trigger_id(),
                              EventFilter :: khepri_evf:event_filter_or_compat(),
                              Action ::
                                  StoredProcPath | MFA | Pid | khepri_event_handler:trigger_action(),
                              StoredProcPath :: khepri_path:path(),
                              MFA :: {module(), atom(), list()},
                              Pid :: pid(),
                              Options :: khepri:command_options() | khepri:trigger_options(),
                              Ret :: ok | khepri:error().

Registers a trigger.

transaction(StoreId, FunOrPath, Args, ReadWrite, Options)

-spec transaction(StoreId, FunOrPath, Args, ReadWrite, Options) -> Ret
                     when
                         StoreId :: khepri:store_id(),
                         FunOrPath :: Fun | PathPattern,
                         Fun :: khepri_tx:tx_fun(),
                         PathPattern :: khepri_path:pattern(),
                         Args :: list(),
                         ReadWrite :: ro | rw | auto,
                         Options :: khepri:command_options() | khepri:query_options(),
                         Ret :: khepri_machine:tx_ret() | khepri_machine:async_ret().

Runs a transaction and returns the result.

unregister_projections(StoreId, Names, Options)

-spec unregister_projections(StoreId, Names, Options) -> Ret
                                when
                                    StoreId :: khepri:store_id(),
                                    Names :: all | [khepri_projection:name()],
                                    Options :: khepri:command_options(),
                                    Ret :: khepri:ok(khepri_machine:projection_map()) | khepri:error().

Removes the given projections from the store.

Names may either be a list of projection names to remove or the atom all. When all is passed, every projection in the store is removed.

version()

-spec version() -> MacVer when MacVer :: 4.

Returns the state machine version.

which_module(MacVer)

-spec which_module(MacVer) -> Module when MacVer :: 0..4, Module :: khepri_machine.

Returns the state machine module corresponding to the given version.