khepri_payload (khepri v0.19.0)

Khepri payloads.

Payloads are the structure used to attach something to a tree node in the store. Khepri supports the following payloads:

No payload at all (no_payload()

Data payload used to store any Erlang term (data())

Stored procedure payload used to store functions (sproc())

Usually, there is no need to explicitly use this module as the type of payload will be autodetected, thanks to the wrap/1 function already called internally.

Types

data/0

-type data() :: #p_data{data :: khepri:data()}.

Internal structure to wrap any Erlang term before it can be stored in a tree node.

The only constraint is the conversion to an Erlang binary must be supported by this term.

no_payload/0

-type no_payload() :: '$__NO_PAYLOAD__'.

Internal value used to mark that a tree node has no payload attached.

payload/0

-type payload() :: no_payload() | data() | sproc().

All types of payload stored in the nodes of the tree structure.

Beside the absence of payload, the only type of payload supported is data.

sproc/0

-type sproc() :: #p_sproc{sproc :: horus:horus_fun(), is_valid_as_tx_fun :: ro | rw | false}.

Internal structure to wrap an anonymous function before it can be stored in a tree node and later executed.

Functions

data(Term)

-spec data(Term) -> Payload when Term :: khepri:data(), Payload :: data().

Returns the same term wrapped into an internal structure ready to be stored in the tree.

See also: data().

none()

-spec none() -> no_payload().

Returns the internal value used to mark that a tree node has no payload attached.

See also: no_payload().

sproc(Fun)

-spec sproc(Fun) -> Payload when Fun :: horus:horus_fun() | fun(), Payload :: sproc().

Returns the same function wrapped into an internal structure ready to be stored in the tree.

See also: sproc().

wrap(Payload)

-spec wrap(Payload) -> WrappedPayload
              when Payload :: payload() | khepri:data() | fun(), WrappedPayload :: payload().

Automatically detects the payload type and ensures it is wrapped in one of the internal types.

The internal types make sure we avoid any collision between any user-provided terms and internal structures.